refactor(intents): migrate capture_event intent to subdirectory architecture
Move CaptureEventIntent from two flat files (1,803 lines) in intents/ to a
new intents/capture_event/ subdirectory package with 9 source files following
the established pattern from browsetimeline/ and burst_management/.

Structural changes:
- Split into constants.go, context.go, result.go, messages.go, interfaces.go,
  types.go, intent.go, handlers.go, helpers.go
- Renamed types to package-scoped names (CaptureEventIntent -> Intent, etc.)
- Moved 9 test files with updated package declarations
- Created capture_event_suite_test.go, benchmarks_test.go, global_keys_test.go
- Updated all consumers (app/, testutil/e2e/, intents/ shared tests)
- Extracted CaptureEvent entries from shared intents/ test files

No backward compatibility layer due to import cycle constraint (capture_event
imports intents for BaseIntent, so intents cannot import capture_event).

AI-Generated-By: OpenCode (Claude Sonnet 4)
