+++ internal/cli/intents/capture_event/burst_extraction_test.go
@@ -1,4 +1,4 @@
-package intents_test
+package capture_event_test
 
 import (
 	"context"
