+++ internal/cli/intents/capture_event/workflow_e2e_test.go
@@ -1,4 +1,4 @@
-package intents_test
+package capture_event_test
 
 import (
 	"strings"
